\name{td.call.sp}
\alias{td.call.sp}
\title{
Locate and call stored procedure
}
\description{
Function searches the Teradata system for stored procedures.
}
\usage{
td.call.sp()
}
\details{
This function allows you to find stored procedures on your Teradata system and
run them via a console menu.  You first search for the stored procedure you are
looking for, select it, enter input parameters if any, and then it is called
for you.
}
\value{
No output results
}
\examples{

\dontrun{
td.call.sp()  #interactive console menu and data entry
}

}
